# Add holiday caching for performance improvement