## Summary

- Implement month-based holiday caching in Calendar class
- Replace per-day API calls with bulk holiday retrieval
- Add error handling for holiday API failures
- Add comprehensive benchmark scripts for performance testing
- Exclude benchmark directory from RuboCop checks

## Performance Results

Benchmark testing shows significant performance improvements:

**Month Display (100 iterations):**
- US: 87.2ms → 1.88ms (46.28x faster)
- JP: 133.81ms → 2.31ms (57.9x faster)  
- GB: 123.34ms → 1.68ms (73.55x faster)

**Year Display (10 iterations):**
- US: 78.04ms → 6.48ms (12.04x faster)
- JP: 201.15ms → 10.88ms (18.49x faster)
- GB: 151.81ms → 9.41ms (16.13x faster)

<details>
<summary>Full Benchmark Results</summary>

```
Fasti Holiday Cache Performance Benchmark
==================================================

Country: US
--------------------
Month Display (July 2024, 100 iterations):
  Legacy (no cache): 87.2ms
  Cached:            1.88ms
  Improvement:       46.28x faster

Year Display (2024, 10 iterations):
  Legacy (no cache): 78.04ms
  Cached:            6.48ms
  Improvement:       12.04x faster


Country: JP
--------------------
Month Display (July 2024, 100 iterations):
  Legacy (no cache): 133.81ms
  Cached:            2.31ms
  Improvement:       57.9x faster

Year Display (2024, 10 iterations):
  Legacy (no cache): 201.15ms
  Cached:            10.88ms
  Improvement:       18.49x faster


Country: GB
--------------------
Month Display (July 2024, 100 iterations):
  Legacy (no cache): 123.34ms
  Cached:            1.68ms
  Improvement:       73.55x faster

Year Display (2024, 10 iterations):
  Legacy (no cache): 151.81ms
  Cached:            9.41ms
  Improvement:       16.13x faster


Benchmark completed!
```

</details>

## Test Plan

- [x] All existing tests pass (154 examples, 0 failures)
- [x] RuboCop checks pass with no violations
- [x] Performance benchmarks demonstrate improvements
- [x] Error handling tested for invalid country codes
- [x] Memory usage remains efficient
- [x] Backward compatibility maintained

## Technical Details

- Uses `Holidays.between()` for bulk retrieval instead of per-day `Holidays.on()` calls
- Implements memoization pattern with proper variable naming (`@holidays_for_month`)
- Pure Ruby implementation without ActiveSupport dependencies
- Graceful error handling with warnings for API failures

:robot: Generated with [Claude Code](https://claude.ai/code)